Move file access support factory.
